requirements.txt
absl-py==0.3.0
astor==0.7.1
certifi==2018.8.13
chardet==3.0.4
future==0.16.0
gast==0.2.0
grpcio==1.14.1
gym==0.10.5
idna==2.7
Markdown==2.6.11
numpy==1.14.5
protobuf==3.6.1
pyglet==1.3.2
requests==2.19.1
six==1.11.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wincertstore==0.2
